backend/requirements.txt
celery==5.4.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
gunicorn==21.2.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8